Refactored peer conductor classes.
