+++ Plugins/NativeServerPlugin/inc/directx_peer_conductor.h
@@ -2,19 +2,19 @@
 
 #include "pch.h"
 #include "peer_conductor.h"
+#include "directx_buffer_capturer.h"
 
 // A PeerConductor using a DirectXBufferCapturer
-class DirectXPeerConductor : public PeerConductor,
-	public has_slots<>
+class DirectXPeerConductor : public PeerConductor
 {
 public:
 	DirectXPeerConductor(int id,
 		const string& name,
-		shared_ptr<WebRTCConfig> webrtcConfig,
-		scoped_refptr<PeerConnectionFactoryInterface> peerFactory,
-		const function<void(const string&)>& sendFunc,
-		ID3D11Device* d3dDevice,
-		bool enableSoftware = false);
+		shared_ptr<WebRTCConfig> webrtc_config,
+		scoped_refptr<PeerConnectionFactoryInterface> peer_factory,
+		const function<void(const string&)>& send_func,
+		ID3D11Device* d3d_device,
+		bool enable_software_encoder = false);
 
 	void SendFrame(ID3D11Texture2D* frame_buffer);
 
@@ -23,9 +23,7 @@ protected:
 	virtual unique_ptr<cricket::VideoCapturer> AllocateVideoCapturer() override;
 
 private:
-	ID3D11Device * m_d3dDevice;
-	bool m_enableSoftware;
-	vector<DirectXBufferCapturer*> m_spies;
-
-	void OnAllocatedPtrDeleted(BufferCapturer* ptr);
+	ID3D11Device* d3d_device_;
+	bool enable_software_encoder_;
+	DirectXBufferCapturer* capturer_;
 };
